feat: rename crates as per minimal one, build new binaries to bucket
