+++ .minimal/minimal.toml
@@ -21,8 +21,8 @@ exec = "cargo build --verbose"
 inherit_cwd = true
 
 [tasks.build-smoke]
-description = "Smoke-test that `minimal run` can build the CLI"
-exec = "cargo build -p minimal"
+description = "Smoke-test that `mip run` can build the CLI"
+exec = "cargo build -p mip"
 inherit_cwd = true
 
 [tasks.test]
