+++ Cargo.toml
@@ -14,8 +14,8 @@ members = [
     "crates/ot",
     "crates/orchestrator",
     "crates/paths",
+    "crates/mip",
     "crates/minimal",
-    "crates/minimal2",
     "crates/minimald",
     "crates/minimald-rpc",
     "crates/minvmd",
