+++ crates/minimal/src/autospawn.rs
