+++ crates/minimal/src/client.rs
