--- crates/minimal2/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "minimal2"
-version = "0.0.1"
-edition.workspace = true
-publish.workspace = true
-
-[dependencies]
-clap.workspace = true
-clap_complete.workspace = true
-camino.workspace = true
-chrono.workspace = true
-dirs.workspace = true
-minimald-rpc.workspace = true
-paths.workspace = true
-russh.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-sessions.workspace = true
-tokio.workspace = true
-tracing.workspace = true
-tracing-subscriber.workspace = true
-
-minvmd = { path = "../minvmd" }
-ot.workspace = true
