+++ crates/mip/src/cmd_add.rs
