+++ crates/mip/src/cmd_cache.rs
