+++ crates/mip/src/cmd_check.rs
