+++ crates/mip/src/cmd_dump.rs
