+++ crates/mip/src/cmd_materialize.rs
