+++ crates/mip/src/cmd_run.rs
